pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "apifrom"
version = "1.0.0"
description = "Transform any Python function into a REST API endpoint"
readme = "README.md"
authors = [
    {name = "sc4rfurry", email = "akalucifr@protonmail.ch"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
keywords = ["api", "rest", "web", "async", "framework"]
dependencies = [
    "starlette>=0.20.0",
    "pydantic>=1.8.0",
    "uvicorn>=0.17.0",
    "typing-extensions>=4.0.0",
    "python-multipart>=0.0.5",
    "jinja2>=3.0.0",
    "httpx>=0.22.0",
    "pyjwt>=2.4.0",
    "cryptography>=37.0.0",
    "passlib>=1.7.4",
]
requires-python = ">=3.7"

[project.optional-dependencies]
database = [
    "sqlalchemy>=1.4.0",
    "asyncpg>=0.25.0",
    "aiomysql>=0.1.0",
    "aiosqlite>=0.17.0",
]
cache = [
    "redis>=4.3.0",
    "aioredis>=2.0.0",
]
monitoring = [
    "prometheus-client>=0.14.0",
]
dev = [
    "pytest>=7.1.0",
    "pytest-asyncio>=0.18.0",
    "pytest-cov>=3.0.0",
    "black>=22.3.0",
    "isort>=5.10.0",
    "mypy>=0.950",
    "flake8>=4.0.0",
]
all = [
    "sqlalchemy>=1.4.0",
    "asyncpg>=0.25.0",
    "aiomysql>=0.1.0",
    "aiosqlite>=0.17.0",
    "redis>=4.3.0",
    "aioredis>=2.0.0",
    "prometheus-client>=0.14.0",
]

[project.urls]
"Homepage" = "https://github.com/sc4rfurry/apifrom"
"Bug Tracker" = "https://github.com/sc4rfurry/apifrom/issues"
"Documentation" = "https://github.com/sc4rfurry/apifrom/blob/main/README.md"
"Source Code" = "https://github.com/sc4rfurry/apifrom" 